server/requirements.txt
pygls==0.13.0
antlr4-python3-runtime==4.13.1
